finalize update users module

+++ src/app/shared/services/users.service.ts
@@ -6,6 +6,7 @@ import { BlockUI, NgBlockUI } from 'ng-block-ui';
 
 import { AuthService } from './auth.service';
 import { ToastService } from './toaster.service';
+import { ModalDataService } from './modal.service';
 
 
 @Injectable()
@@ -19,19 +20,19 @@ export class UsersService {
     constructor(
         private httpClient: HttpClient,
         private auth: AuthService, 
-        private toastService: ToastService) { 
+        private toastService: ToastService,
+        private mdService: ModalDataService) { 
     }
 
-    // base apis url
-    private baseUrl = 'https://reqres.in/api';
 
     // get all users request
     getUsers() {
         this.blockUI.start();
-        return this.httpClient.get(this.baseUrl + '/users', { observe: 'response', 'headers': this.auth.headers })
+        return this.httpClient.get(this.auth.baseUrl + '/users', { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError(error.message || 'Server Error')
-                return Observable.throw(error.message || 'Server Error')
+                this.mdService.closeDialog();
+                this.toastService.showError(error.message || 'Server Error');
+                return Observable.throw(error.message || 'Server Error');
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -45,10 +46,11 @@ export class UsersService {
     // get single user request
     getUser(id: Number) {
         this.blockUI.start();
-        return this.httpClient.get(this.baseUrl + "/users/" + id, { observe: 'response', 'headers': this.auth.headers })
+        return this.httpClient.get(this.auth.baseUrl + '/users/' + id, { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError(error.message || 'Server Error')
-                return Observable.throw(error.message || 'Server Error')
+                this.mdService.closeDialog();
+                this.toastService.showError(error.message || 'Server Error');
+                return Observable.throw(error.message || 'Server Error');
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -62,10 +64,11 @@ export class UsersService {
     // create single user request
     createUser(data: any) {
         this.blockUI.start();
-        return this.httpClient.post(this.baseUrl + '/users', data, { observe: 'response', 'headers': this.auth.headers })
+        return this.httpClient.post(this.auth.baseUrl + '/users', data, { observe: 'response', 'headers': this.auth.headers })
             .catch((error: any) => {
-                this.toastService.showError(error.message || 'Server Error')
-                return Observable.throw(error.message || 'Server Error')
+                this.mdService.closeDialog();
+                this.toastService.showError(error.message || 'Server Error');
+                return Observable.throw(error.message || 'Server Error');
             })
             .finally(() => {
                 this.pendingRequests--;
@@ -76,28 +79,31 @@ export class UsersService {
             }) as any;
     }
 
-    // updateUser(id: any, data: any, {}) {
-    //     this.blockUI.start();
-    //     return this.httpClient.delete(this.baseUrl + '/users/' + id, data, { 'headers': this.authInterceptor.headers })
-    //         .catch((error: any) => {
-    //             this.toastService.showError("User can't be deleted!")
-    //             return Observable.throw(error);
-    //         })
-    //         .finally(() => {
-    //             this.pendingRequests--;
-    //             if (this.pendingRequests <= 0) {
-    //                 // there are no more pending requests
-    //                 this.blockUI.stop();
-    //             }
-    //         }) as any;
-    // }
+    // update single user data
+    updateUser(id: any, data: any) {
+        this.blockUI.start();
+        return this.httpClient.put(this.auth.baseUrl + '/users/' + id, data, { observe: 'response', 'headers': this.auth.headers })
+            .catch((error: any) => {
+                this.mdService.closeDialog();
+                this.toastService.showError(error.message || 'Server Error');
+                return Observable.throw(error.message || 'Server Error');
+            })
+            .finally(() => {
+                this.pendingRequests--;
+                if (this.pendingRequests <= 0) {
+                    // there are no more pending requests
+                    this.blockUI.stop();
+                }
+            }) as any;
+    }
 
     // delete single user request
     deleteUser(id: any) {
         this.blockUI.start();
-        return this.httpClient.delete(this.baseUrl + '/users/' + id, { observe: 'response', 'headers': this.auth.headers })  
+        return this.httpClient.delete(this.auth.baseUrl + '/users/' + id, { observe: 'response', 'headers': this.auth.headers })  
             .catch((error: any) => {
-                this.toastService.showError(error.message || 'Server Error')
+                this.mdService.closeDialog();
+                this.toastService.showError(error.message || 'Server Error');
                 return Observable.throw(error.message || 'Server Error');
             })
             .finally(() => {
